+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2289 - can change participants & slot directly from admin meeting edition
+
 ### Fixed
 - MV-313 - trunck happening name if too long for tokbox
 - Code style / namespace fixes